BUG: Workaround a segfault building TVTK on VTK5.9
In the TVTK building process, to associate numerical values with method
names in setters, these setters are called. For the
vtk.vtkResliceImageViewer, calling the setter 'SetResliceModeToOblique'
causes the Image's ComputeBounds method to be called. This triggers a
segfault, as the image is not allocated. We work around this segfault
by not doing this inspection for viewer objects (namely
vtk.vtkImageViewer and vtk.vtkResliceImageViewer)