Allow setting a list of contours for a contour surface
The surface module sets auto_contours to True by default, as opposed to
the iso_surface module which has it set to False.
This meant that passing a list to the contours argument when creating a
contour surface did not work unless auto_contours was also set to False.
To fix this, when setting a list of contour values, make sure
auto_contours is set to False so that the list is used.