BUG: Qt: fix crashing full-screen-mode.
Previously, when you click on full-screen and press escape/q/e it would
segfault.  The earlier full screen would also not allow the user
complete interaction with VTK widgets and the like.  This has been fixed
although the behavior of the full-screen is different -- it basically
full-screen's the entire application.  On an mlab window the behaviour
will be almost as before.  This turns out to be quite useful in
practice.