The initial alpha version of a test to check memory usage and leaks.
This is cloned from the contour integration test. The test runs the scenario of
adding a scene, configuring it with components, changing the components, doing
a few checks, then it saves, closes the scene and restores it.

As of this commit, it reports the memory leaks per sample run and for all runs.
When the leaks are fully fixed, the leak should tend to ZERO!