Allow users to import tvtk without an X server.
Currently this is not possible since there are imports to traitsui.api.
For example if one is connected to a machine via ssh and no X11
forwarded and simply does `from tvtk.api import tvtk`, the application
will close.  This should allow a user to use TVTK without a UI/X server
just as they would with VTK.