Fix many of the Mayavi objects.
Most of them use the new pipeline now and the connections should be set
correctly for many of the objects.  All the tests pass with these
changes.  There is also a new convenient method called
`get_output_dataset` that can be used when one needs the output.  This
should be cleaned up later.