Updated LUTs to include new matplotlib colorschemes
I had to modify the cm2lut.py script to get the cmaps which are not in
matplotlib.cm.datad. I don't understand why some of the cmaps are there
and others are in matplotlib._cm_listed.cmaps, but I figured out from
matplotlib.cm.get_cmap that they are.