First cut of notebook support.
There are two modes supported, one is to dump PNG images for the scene
and the second (the default) is to use x3d output and x3dom.js to
render the scene.  This works very well.  This also adds the x3dom.js
file to Mayavi so users can view the data even when disconnected from
the internet.