Fix issue with vmin/vmax for pipeline.volume.
There was a test failure with `mlab.pipeline.volume`'s vmin and vmax
traits.  This happens because for some reason in the past the rescaling
seemed to be called once when both vmin and vmax were correctly set.  It
is not clear how this happened, perhaps a change to traits or perhaps a
Python version change as only the handler for _vmin_changed was being
used.  Currently though, this is not guaranteed and vmax is None when
both vmin and vmax are passed and the first time the range is scaled.
So when the values were rescaled they used the wrong range.  Now we
simply store the last min/max used for the value scaling and this fixes
the problem.