BUG: fix subtle windows specific error.
The tests were failing because on Windows alone we were unable to import
image_connectivity_filter module from TVTK's zipped classes.  This in
turn happened because one of the traits there was declared effectively
as,
`size_range = traits.Array(dtype=int, value=(1, 9223372036854775807))`
Now on *nix, the dtype of int maps to int64 whereas on windows the
default C integer defaults to int32.  Therefore this was an invalid
value for the defined range and the trait was invalid causing all sorts
of errors.  This is now fixed by using int64 explicitly on windows.