Fix a bug with restoring a scene.
When a scene is restored from saved state using the state pickler, the
light manager may not be ready yet as the widget may not be painted.
This causes test failures and also the tests to pop up a dialog.  This
is now handled correctly by deferring loading the state to a trait
handler which looks for the saved state.