Make array_ext optional.
On a Python environment where compilation is not possible, this
extension prevents installation via pip.  We now have a slower
alternative for such environments which allow us to continue to install
Mayavi without building the extension.  The extension is only added
conditionally, so those who can build the package correctly will get the
added performance.  This addresses #658 and should make Mayavi even
easier to install for the average user.