BUG: wrap the `Get` methods correctly.
Many of the new pipeline methods were not wrapped correctly.  For
example the vtkAlgorithm.GetInputAlgorithm has multiple signatures.
TVTK was wrapping any getter which had one of its signatures with no
args as a pure property and not exposing the method itself.  This means
that users cannot call `obj.get_input_algorithm(0, 0)` which is broken.
We now wrap the no arg call as a property but also wrap the generic
method as a callable method.