Generalize the pipeline browser.
It can now be used anywhere.  The UI now shows a dialog for the object
in a split pane below the tree.  A new event object_edited fires when
any of the objects has been edited on the UI.  The view is now generated
in a default_traits_view allowing us to embed the pipeline browser
elsewhere.