Allow adding any tvtk.Object to pipeline.
This is convenient as a user can also create a full visualization and
then simply add the actor to the pipeline via
`mlab.pipeline.add_dataset(actor)`.  If the user simply appends this
actor to the resulting source's actor list it will also be rendered on
the scene.  This is handy when something is doable with VTK but not with
mayavi.