Add powerful new jupyter notebook backend.
This backend is called 'ipy' and is now the default.  It requires
that VTK's offscreen support work correctly.  This backend requires
ipywidgets and ipyevents.  It allows complete interactivity and behaves
almost exactly like a normal UI backend but inside a notebook.  Note
that this will still require an xserver or windowing toolkit unless VTK
is compiled to be able to work without those.