Fix issue #770.
The issue arises in a few of the mlab sources where when we reset to a
polydata with a smaller number of points, we first set the points and
the original polys are still pointing to the old points which no longer
exist.  We now first set the polys to None before changing anything.