BUG: Fix bug with picker.
When picking, there is no way to remove the picker UI except picking
away from an actor.  However, when this happens an exception was raised
due to a silly bug.  This also changes the behavior now so we press
escape/q/e keys it will no longer display the picker UI.  Also fixes a
bug with the wxPython scene for VTK 9.x.