ENH: Load tvtk class names dynamically.
The `get_tvtk_class_names` function instantiates all the VTK classes
at import time.  This reduces import time and is wasteful.  We now do
this only when required.  This is also very likely the reason we run
into crashes when VTK versions change and have some buggy classes which
we do not even use.