Fix issue with `mlab.process_ui_events`
The method was basically doing nothing when no UI toolkit was explicitly
set. Since this import is early it would so happen that the toolkit was
not set at this point. As a result `mlab.process_ui_events` was doing
nothing. In keeping with the Zen of Python, we now do not import pyface
only when the toolkit is explicitly set to 'null' or the CI env var is
set.